+++ src/Resources/AbstractUnzerResource.php
@@ -525,17 +525,16 @@ abstract class AbstractUnzerResource implements UnzerParentInterface
 
             // handle additionalTransactionData values
             if ($property === 'additionalTransactionData') {
-                if (!empty($value->riskData) && $value->riskData instanceof self) {
-                    $value->riskData = $value->riskData->expose();
-                }
-                if (!empty($value->shipping) && $value->shipping instanceof self) {
-                    $value->shipping = $value->shipping->expose();
+                foreach ($value as $key => $data) {
+                    if ($data instanceof self) {
+                        $value->$key = $data->expose();
+                    }
                 }
             }
 
             $properties[$property] = $value;
         }
-    
+
         return $properties;
     }
 
@@ -552,7 +551,10 @@ abstract class AbstractUnzerResource implements UnzerParentInterface
          * @var AbstractUnzerResource $linkedResource
          */
         foreach ($this->getLinkedResources() as $attributeName => $linkedResource) {
-            $exposedResources[$attributeName . 'Id'] = $linkedResource ? $linkedResource->getId() : '';
+            $resourceId = $linkedResource ? $linkedResource->getId() : null;
+            if ($resourceId !== null) {
+                $exposedResources[$attributeName . 'Id'] = $resourceId;
+            }
         }
         return $exposedResources;
     }