+++ src/Unzer.php
@@ -729,6 +729,11 @@ class Unzer implements UnzerParentInterface, PaymentServiceInterface, ResourceSe
         return $this->paymentService->performAuthorization($authorization, $paymentType, $customer, $metadata, $basket);
     }
 
+    public function updateAuthorization($payment, Authorization $authorization): Authorization
+    {
+        return $this->paymentService->updateAuthorization($payment, $authorization);
+    }
+
     /**
      * {@inheritDoc}
      */
@@ -772,6 +777,11 @@ class Unzer implements UnzerParentInterface, PaymentServiceInterface, ResourceSe
     public function performCharge(Charge $charge, $paymentType, $customer = null, $metadata = null, $basket = null): Charge
     {
         return $this->paymentService->performCharge($charge, $paymentType, $customer, $metadata, $basket);
+    }
+
+    public function updateCharge($payment, Charge $charge): Charge
+    {
+        return $this->paymentService->updateCharge($payment, $charge);
     }
 
     /**